reader/requirements.txt
requests==2.31.0
gtfs-realtime-bindings==1.0.0
google-cloud-pubsub==2.27.1